LibGDX: Cleaned a bit (part 6)

+++ libgdx/tests/gdx-tests/src/com/badlogic/gdx/tests/g3d/FrameBufferCubemapTest.java
@@ -92,16 +92,16 @@ public class FrameBufferCubemapTest extends Basic3DSceneTest {
     }
 
     public void renderCube() {
-        int w = Gdx.graphics.getBackBufferWidth();
-        int h = Gdx.graphics.getBackBufferHeight();
+        float w = Gdx.graphics.getBackBufferWidth();
+        float h = Gdx.graphics.getBackBufferHeight();
         int x = (int) (w - w * 0.5f);
         int y = (int) (h - h * 0.5f);
         w *= 0.5f;
         h *= 0.5f;
 
-        Gdx.gl.glViewport(x, y, w, h);
+        Gdx.gl.glViewport(x, y, (int) w, (int) h);
         Gdx.gl.glEnable(GL20.GL_SCISSOR_TEST);
-        Gdx.gl.glScissor(x, y, w, h);
+        Gdx.gl.glScissor(x, y, (int) w, (int) h);
         Gdx.gl.glClearColor(1, 1, 1, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT | GL20.GL_DEPTH_BUFFER_BIT);
